+++ util/tls/CMakeLists.txt
@@ -1,6 +1,6 @@
 Message(STATUS "OpenSSL libs ${OPENSSL_SSL_LIBRARIES} ${OPENSSL_VERSION}")
 
-add_library(tls_lib tls_engine.cc tls_socket.cc iovec_utils.cc)
+add_library(tls_lib tls_engine.cc tls_socket.cc tls_socket_async.cc iovec_utils.cc)
 
 cxx_link(tls_lib fibers2 OpenSSL::SSL)
 helio_cxx_test(tls_engine_test tls_lib LABELS CI)
